@@ -24,25 +24,25 @@
     "packages/common/testing"
   ],
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.125",
-    "@types/jest": "^29.5.6",
-    "@types/node": "^18.11.4",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.9.0",
+    "@types/aws-lambda": "^8.10.130",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.10.3",
+    "@typescript-eslint/eslint-plugin": "^6.13.2",
+    "@typescript-eslint/parser": "^6.13.2",
     "aws-lambda": "^1.0.7",
-    "eslint": "^8.52.0",
-    "eslint-config-prettier": "^9.0.0",
+    "eslint": "^8.55.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import-newlines": "^1.3.4",
     "eslint-plugin-prettier": "^5.0.1",
     "jest": "^29.7.0",
     "jest-junit": "^16.0.0",
     "license-checker": "^25.0.1",
-    "prettier": "^3.0.3",
+    "prettier": "^3.1.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.2"
   },
   "dependencies": {
-    "esbuild": "^0.19.5"
+    "esbuild": "^0.19.8"
   }
 }
