@@ -1,9 +1,13 @@
 import {APIGatewayProxyEvent, APIGatewayProxyResult} from "aws-lambda"
 import {handler} from "../src/getMyPrescriptions"
-import {expect, describe, it} from "@jest/globals"
+import {
+  expect,
+  describe,
+  it,
+  jest
+} from "@jest/globals"
 import {ContextExamples} from "@aws-lambda-powertools/commons"
 import {Logger} from "@aws-lambda-powertools/logger"
-import "jest"
 import MockAdapter from "axios-mock-adapter"
 import axios from "axios"
 import {mockAPIGatewayProxyEvent} from "@prescriptionsforpatients_common/testing"
