@@ -6,6 +6,7 @@ import {Agent} from "https"
 import {Logger} from "@aws-lambda-powertools/logger"
 
 const mock = new MockAdapter(axios)
+const axiosInstance = axios.create()
 
 describe("Health check", () => {
   const logger = new Logger({serviceName: "spineClient"})
@@ -18,7 +19,7 @@ describe("Health check", () => {
   test("Successful health check result returns success", async () => {
     mock.onGet("/healthcheck").reply(200, {})
 
-    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent)
+    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent, axiosInstance)
 
     expect(spineResponse.status).toBe("pass")
     expect(spineResponse.responseCode).toBe(200)
@@ -29,7 +30,7 @@ describe("Health check", () => {
   test("Failure health check result returns failure", async () => {
     mock.onGet("/healthcheck").reply(500, {})
 
-    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent)
+    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent, axiosInstance)
 
     expect(spineResponse.status).toBe("error")
     expect(spineResponse.responseCode).toBe(500)
@@ -40,7 +41,7 @@ describe("Health check", () => {
   test("health check network issues result returns failure", async () => {
     mock.onGet("/healthcheck").networkError()
 
-    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent)
+    const spineResponse = await serviceHealthCheck("healthcheck", logger, httpsAgent, axiosInstance)
 
     expect(spineResponse.status).toBe("error")
     expect(spineResponse.responseCode).toBe(500)
